# Conflicts:
#	server/catgenome/src/main/java/com/epam/catgenome/manager/blast/BlastTaskManager.java

+++ server/catgenome/src/main/java/com/epam/catgenome/manager/blast/dto/BlastRequest.java
@@ -15,6 +15,6 @@ public class BlastRequest {
     private List<Long> excludedTaxIds;
     private String query;
     private Long maxTargetSequence;
-    private Long expectedThreshold;
+    private Double expectedThreshold;
     private String options;
 }
